Move getTimeNotification to utils
