+++ frontend/src/components/details/CountDown.js
@@ -1,6 +1,7 @@
 import React, { Component, Fragment } from "react"
 import "../../css/Reuse.scss" //도구성컴포넌트의CSS(SCSS)
 import "../../App.css" //공통CSS
+import { getTimeNotification } from "../../modules/utils"
 
 class CountDown extends Component {
   constructor(props) {
@@ -44,56 +45,12 @@ class CountDown extends Component {
     this.ifTimer && clearInterval(this.ifTimer)
   }
 
-  timeNotification = seconds => {
-    let gapDatePart = parseInt(seconds / (60 * 60 * 24))
-    let gapHourPart = parseInt(
-      (seconds - gapDatePart * 60 * 60 * 24) / (60 * 60)
-    )
-    let gapMinutePart = parseInt(
-      (seconds - gapDatePart * 60 * 60 * 24 - gapHourPart * 60 * 60) / 60
-    )
-    let gapSecondPart =
-      seconds -
-      gapDatePart * 60 * 60 * 24 -
-      gapHourPart * 60 * 60 -
-      gapMinutePart * 60
-
-    if (gapHourPart < 10 && gapDatePart !== 0) {
-      gapHourPart = "0" + gapHourPart
-    }
-
-    if (gapMinutePart < 10) {
-      gapMinutePart = "0" + gapMinutePart
-    }
-
-    if (gapSecondPart < 10) {
-      gapSecondPart = "0" + gapSecondPart
-    }
-
-    if (gapDatePart < 1) {
-      return (
-        "남은시간 " + gapHourPart + ":" + gapMinutePart + ":" + gapSecondPart
-      )
-    } else {
-      return (
-        "남은시간 " +
-        gapDatePart +
-        "일 " +
-        gapHourPart +
-        ":" +
-        gapMinutePart +
-        ":" +
-        gapSecondPart
-      )
-    }
-  }
-
   render() {
     const { gapSecondTotal } = this.state
 
     return (
       <Fragment>
-        {gapSecondTotal > 0 && this.timeNotification(gapSecondTotal)}
+        {gapSecondTotal > 0 && getTimeNotification(gapSecondTotal)}
       </Fragment>
     )
   }
